Make EPI field read only after the cluster is created
