+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/platformIntegration/constants.tsx
@@ -1,0 +1,19 @@
+import { NUTANIX_CONFIG_LINK, VSPHERE_CONFIG_LINK } from '../../../../common';
+import { PlatformType } from '../../../api';
+
+export const ExternalPlatformLabels: { [key in PlatformType]: string } = {
+  baremetal: 'No platform integration',
+  none: 'No platform integration',
+  nutanix: 'Nutanix',
+  oci: 'Oracle  (Requires a custom manifest)',
+  vsphere: 'vSphere',
+};
+
+export const ExternalPlatformLinks: Partial<{ [key in PlatformType]: string }> = {
+  nutanix: NUTANIX_CONFIG_LINK,
+  vsphere: VSPHERE_CONFIG_LINK,
+};
+
+export const ExternalPlatformTooltips: Partial<{ [key in PlatformType]: string }> = {
+  oci: "To integrate with an external partner (for example, Oracle Cloud), you'll need to provide a custom manifest.",
+};
