+++ libs/ui-lib/lib/ocm/components/clusterWizard/ClusterDetailsForm.tsx
@@ -184,7 +184,6 @@ const ClusterDetailsForm = (props: ClusterDetailsFormProps) => {
                   clusterExists={!!cluster}
                   clusterCpuArchitecture={cluster?.cpuArchitecture as CpuArchitecture}
                   clusterId={cluster?.id}
-                  clusterPlatform={cluster?.platform?.type}
                 />
               </GridItem>
             </Grid>
